+++ R/TargetingModels.R
@@ -413,8 +413,12 @@ createSubstitutionMatrix <- function(db, model=c("RS", "S"),
 #'                               to compute the mutability rates. If the number is smaller 
 #'                               than this threshold, the mutability for the 5-mer will be 
 #'                               inferred. Default is 500.    
+#' @param    numSeqMutationsOnly return only a vector counting the observed number of mutations 
+#'                               in sequences containing each 5-mer. This option can be used for
+#'                               parameter tuning for \code{minNumSeqMutations} during 
+#'                               preliminary analysis. Default is \code{FALSE}.                              
 #' @param    returnSource        return the sources of 5-mer mutabilities (measured vs.
-#'                               inferred). Default is false.                          
+#'                               inferred). Default is \code{FALSE}.                          
 #'
 #' @return   A named numeric vector of 1024 normalized mutability rates for each 5-mer 
 #'           motif with names defining the 5-mer nucleotide sequence.
@@ -445,6 +449,7 @@ createMutabilityMatrix <- function(db, substitutionModel, model=c("RS", "S"),
                                    vCallColumn="V_CALL",
                                    multipleMutation=c("independent", "ignore"),
                                    minNumSeqMutations=500, 
+                                   numSeqMutationsOnly=FALSE,
                                    returnSource=FALSE) {
     # substitutionModel=sub_model; model="S"; sequenceColumn="SEQUENCE_IMGT"; germlineColumn="GERMLINE_IMGT_D_MASK"
     # vCallColumn="V_CALL"; multipleMutation="ignore"; minNumSeqMutations=10; returnSource=FALSE
@@ -589,8 +594,12 @@ createMutabilityMatrix <- function(db, substitutionModel, model=c("RS", "S"),
     
     # Filter out 5-mers with low number of observed mutations in the sequences
     NumSeqMutations <- sapply(1:1024,function(i)sum(MutabilityWeights[!is.na(MutabilityMatrix[i,])])) 
+    names(NumSeqMutations) = names(Mutability_Mean)
     Mutability_Mean[NumSeqMutations < minNumSeqMutations] <- NA
     
+    if (numSeqMutationsOnly) {
+        return(NumSeqMutations)
+    }
     
     # Infer mutability for missing 5-mers
     .fillHot <-function(FIVEMER,mutability){
@@ -1137,7 +1146,7 @@ writeTargetingDistance <- function(model, file) {
 #' 
 #' \code{plotMutability} plots the mutability rates of a \code{TargetingModel}.
 #' 
-#' @param    model        \link{TargetingModel} object or matrix containing normalized 
+#' @param    model        \link{TargetingModel} object or vector containing normalized 
 #'                        mutability rates.
 #' @param    nucleotides  vector of center nucleotide characters to plot.
 #' @param    style        type of plot to draw. One of: