+++ NEWS.md
@@ -1,3 +1,18 @@
+Version 0.1.4:  August 5, 2016
+-------------------------------------------------------------------------------
+
+Selection Analysis:
+
++ Fixed a bug in calcBaseline wherein the germline column was incorrected 
+  hardcoded, leading to erroneous mutation counts for some clonal consensus 
+  sequences.
+
+Targeting Models:
+
++ Added `numSeqMutationsOnly` argument to `createMutabilityMatrix()`, enabling
+  parameter tuning for `minNumSeqMutations`.
+  
+
 Version 0.1.3:  July 31, 2016
 -------------------------------------------------------------------------------